bug fixing 3

+++ agent/mailTask.py
@@ -216,7 +216,7 @@ class MailTask(object):
 
         return - True if a job is finished
         """
-        if os.path.isfile(os.path.join(UPLOAD_FOLDER, max(os.listdir(UPLOAD_FOLDER)), RESULT_JSON)):
+        if os.path.isdir(UPLOAD_FOLDER) and os.path.isfile(os.path.join(UPLOAD_FOLDER, max(os.listdir(UPLOAD_FOLDER)), RESULT_JSON)):
             if os.path.isfile(RESULT_ZIP):
                 os.remove(RESULT_ZIP)
             shutil.make_archive("result", 'zip', os.path.join(UPLOAD_FOLDER, max(os.listdir(UPLOAD_FOLDER))))
